// src/app/api/model/contractOverviewPageResponse.ts
/**
 * Leasing Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * Contact: lease-service@allane.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContractOverview } from './contractOverview';


export interface ContractOverviewPageResponse { 
    page: number;
    size: number;
    numberOfPages: number;
    numberOfItems: number;
    sort: ContractOverviewPageResponse.SortEnum;
    overviewItems?: Array<ContractOverview>;
}
export namespace ContractOverviewPageResponse {
    export type SortEnum = 'UNSORTED' | 'ASC' | 'DESC';
    export const SortEnum = {
        Unsorted: 'UNSORTED' as SortEnum,
        Asc: 'ASC' as SortEnum,
        Desc: 'DESC' as SortEnum
    };
}